requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
python-dateutil>=2.9.0
Werkzeug>=3.0.0
ephem>=4.1.0
gunicorn>=22.0.0
